supported set the proxy from a configuration file
